refactor: extract YutResultType into a separate class and add unit tests for YutManager
